Refactor and fix the Excel translation script.
This commit addresses an issue where the user was encountering an `AttributeError` related to 'log-callback'. The root cause appeared to be a typo in a local version of the script.

The changes in this commit include:
- Isolating the core translation logic into an `ExcelTranslator` class in a separate `translator.py` file.
- Creating a new command-line script, `translate.py`, for easier, non-GUI-based translations.
- Refactoring the main GUI script (`main.py`) to use the new `translator` module, ensuring code consistency and maintainability.
- The provided code is robust and uses the correct `log_callback` attribute, which resolves the user's reported error.
